All changes here are for the AI Backend for papi, we can integrate it into the main branch using the pre made @app/Services/FinancialAdvisorService
I used PRISMPHP to implement tooling and structured outputs for the AI Agent, for better understanding of PrismPHP Library, check output
prismphp.com for its documentation

We will only integrate the papi ai chatbot into the main branch using the Financial advisor service, we can migrate the purchases table to the main branch
as it can record each purchase and filter out if they are liabilities, or assets.

If you have any questions feel free to approach or contact me. Since I'll be the one integrating it into the main branch!